fix(popover): allow null and undefined as values for popover and title
This makes it consistent with the behavior of tooltip. It also makes more sense since the title is an optional input, and its default value is thus undefined (which, in strict mode, is neither a string nor a TemplateRef).
Note that a test checked that no title was displayed when the title was empty, but it actually checked that no title was displayed if the title was undefined. And it *did* display an empty title if the title was an empty string.
Now, the title is displayed only is it is truthy.

refs #3845