test(datepicker): ignore warnings that are unavoidable
Track-related warnings now cause test failures by default.
So, if such a failure happens, we can either fix the code to avoid the failure when the code is incorrect (and the error is thus justified), or ignore the warning when the code is correct and the warning is unavoidable.

In the end, the test report is not be spammed with warnings anymore, and when there is a warning, the error allows knowing which test caused it.